requirements.txt
flask==2.3.3
python-telegram-bot==20.7
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0 